security commit

+++ src/etudiant/etudiant.controller.ts
@@ -1,13 +1,15 @@
 
-import { Body, Controller, Delete, Get, Param, Patch, Post, UploadedFile, UseInterceptors } from '@nestjs/common';
+import { Body, Headers, Res, HttpStatus, Controller, Delete, Get, Param, Patch, Post, UploadedFile, UseInterceptors } from '@nestjs/common';
 import { EtudiantService } from './etudiant.service';
 import { FileInterceptor } from '@nestjs/platform-express';
+import { Response } from 'express';
 import { diskStorage } from 'multer';
 import { User } from 'src/user/user.model';
+import { LocalStrategy } from 'src/auth/local.auth';
 
 @Controller('etudiant')
 export class EtudiantController {
-    constructor( private readonly etudiantService : EtudiantService){}
+    constructor( private readonly etudiantService : EtudiantService, private authService: LocalStrategy){}
     @Post()
     @UseInterceptors(FileInterceptor('Resume', {
       storage: diskStorage({
@@ -28,9 +30,21 @@ export class EtudiantController {
       return { id: generatedId, filePath: file.path };
     }
     @Get()
-        async getAllEtudiants(){
-            const etudiants = await this.etudiantService.getEtudiants();
-            return etudiants;
+        async getAllEtudiants(    
+          @Res() res: Response,
+          @Headers('authorization') authHeader: string
+        ){
+          if(authHeader){
+            const token = authHeader.split(' ')[1];
+            const userFound = await this.authService.validateUser(token);
+        
+            if (!userFound)
+              return res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' });
+            else {
+              const etudiants = await this.etudiantService.getEtudiants();
+              return etudiants;
+            }
+          }
         }
     @Post('/upload')
     @UseInterceptors( FileInterceptor('file', {
@@ -49,14 +63,36 @@ export class EtudiantController {
         }
     @Get(':id')
         
-        getEtudiant(@Param('id') id: string) {
-          return this.etudiantService.getSingleEtudiant(id);
+        async getEtudiant(@Param('id') id: string,           @Res() res: Response,
+        @Headers('authorization') authHeader: string) {
+          const token = authHeader.split(' ')[1];
+          const userFound = await this.authService.validateUser(token);
+      
+          if (!userFound)
+            return res
+              .status(HttpStatus.UNAUTHORIZED)
+              .json({ message: 'Unauthorized' });
+          else {
+            return this.etudiantService.getSingleEtudiant(id);
+          }
         }
     @Delete(':id')
-        async removeEtudiant(@Param('id') Id: string){
+        async removeEtudiant(@Param('id') Id: string,           @Res() res: Response,
+        @Headers('authorization') authHeader: string){
+          const token = authHeader.split(' ')[1];
+          const userFound = await this.authService.validateUser(token);
+      
+          if (!userFound)
+            return res
+              .status(HttpStatus.UNAUTHORIZED)
+              .json({ message: 'Unauthorized' });
+          else {
             await this.etudiantService.deleteEtudiant(Id);
             return "Deleted Successfully";
         }
+      }
+
+
         @Patch(':id')
   @UseInterceptors(FileInterceptor('Resume', {
     storage: diskStorage({
@@ -71,9 +107,19 @@ export class EtudiantController {
     @Body('competences') competences: Array<string>,
     @Body('localisation') localisation: string,
     @UploadedFile() file: Express.Multer.File,
+    @Res() res: Response,
+    @Headers('authorization') authHeader: string
   ) {
-    const updatedEtudiant = await this.etudiantService.updateEtudiant(id, competences, localisation, file.path);
-    
-    return updatedEtudiant;
+    const token = authHeader.split(' ')[1];
+    const userFound = await this.authService.validateUser(token);
+
+    if (!userFound)
+      return res
+        .status(HttpStatus.UNAUTHORIZED)
+        .json({ message: 'Unauthorized' });
+    else {
+      const updatedEtudiant = await this.etudiantService.updateEtudiant(id, competences, localisation, file.path);
+      return updatedEtudiant;
+    }
   }
 }
