+++ src/offre-stage/offre-stage.controller.ts
@@ -1,14 +1,17 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post } from '@nestjs/common';
+import { Body, Headers, Res, HttpStatus, Controller, Delete, Get, Param, Patch, Post } from '@nestjs/common';
 import { OffreStageService } from './offre-stage.service';
+import { Response } from 'express';
 import { Etudiant } from 'src/etudiant/etudiant.model';
 import { Entreprise } from 'src/entreprise/entreprise.model';
+import { LocalStrategy } from 'src/auth/local.auth';
 
 @Controller('offre-stage')
 export class OffreStageController {
-    constructor(private readonly offreStageService: OffreStageService) { }
+    constructor(private readonly offreStageService: OffreStageService, private authService: LocalStrategy) { }
     @Post()
     async addOffreStage(
-        
+        @Res() res: Response,
+        @Headers('authorization') authHeader: string,
         @Body('titre') titre: string,
         @Body('description') description: string,
         @Body('domaine') domaine: string,
@@ -19,17 +22,40 @@ export class OffreStageController {
         @Body('date_fin') date_fin: Date,
         @Body('entreprise') entreprise: Entreprise,
     ) {
-        const generatedId = await this.offreStageService.insertOffreStage(titre, description, domaine, candidatures, competences, localisation ,date_debut, date_fin,entreprise);
-        return { id: generatedId };
+        const token = authHeader.split(' ')[1];
+        const userFound = await this.authService.validateUser(token);
+    
+        if (!userFound)
+          return res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' });
+        else {
+            const generatedId = await this.offreStageService.insertOffreStage(titre, description, domaine, candidatures, competences, localisation ,date_debut, date_fin,entreprise);
+            return { id: generatedId };
+        }
     }
     @Get()
-    async getAllOffresStage() {
-        const offresStage = await this.offreStageService.getOffresStage();
-        return offresStage;
+    async getAllOffresStage(          @Res() res: Response,
+    @Headers('authorization') authHeader: string) {
+        const token = authHeader.split(' ')[1];
+        const userFound = await this.authService.validateUser(token);
+    
+        if (!userFound)
+          return res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' });
+        else {
+            const offresStage = await this.offreStageService.getOffresStage();
+            return offresStage;
+        }
     }
     @Get(':id')
-    getOffreStage(@Param('id') id: string) {
-        return this.offreStageService.getSingleOffreStage(id);
+    async getOffreStage(@Param('id') id: string, @Res() res: Response,
+    @Headers('authorization') authHeader: string) {
+        const token = authHeader.split(' ')[1];
+        const userFound = await this.authService.validateUser(token);
+    
+        if (!userFound)
+          return res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' });
+        else {
+            return this.offreStageService.getSingleOffreStage(id);
+        }
     }
     @Patch(':id')
     async updateOffreStage(
@@ -41,14 +67,32 @@ export class OffreStageController {
         @Body('localisation') localisation: string,
         @Body('date_debut') date_debut: Date,
         @Body('date_fin') date_fin: Date,
+        @Res() res: Response,
+        @Headers('authorization') authHeader: string
     ) {
-        await this.offreStageService.updateOffreStage(id, titre, description, domaine, competences, localisation ,date_debut, date_fin);
-        return "Updated Successfully";
+        const token = authHeader.split(' ')[1];
+        const userFound = await this.authService.validateUser(token);
+    
+        if (!userFound)
+          return res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' });
+        else {
+            await this.offreStageService.updateOffreStage(id, titre, description, domaine, competences, localisation ,date_debut, date_fin);
+            return "Updated Successfully";
+        }
     }
     @Delete(':id')
-    async removeOffreStage(@Param('id') Id: string) {
-        await this.offreStageService.deleteOffreStage(Id);
-        return "Deleted Successfully";
+    async removeOffreStage(@Param('id') Id: string,
+    @Res() res: Response,
+    @Headers('authorization') authHeader: string) {
+        const token = authHeader.split(' ')[1];
+        const userFound = await this.authService.validateUser(token);
+    
+        if (!userFound)
+          return res.status(HttpStatus.UNAUTHORIZED).json({ message: 'Unauthorized' });
+        else {
+            await this.offreStageService.deleteOffreStage(Id);
+            return "Deleted Successfully";
+        }
     }
 
 }
